Move `iter_flatten` out of tests
It's a dependency of the in-memory engine I missed in #772
