+++ databuilder/query_engines/in_memory_database.py
@@ -7,8 +7,8 @@ See tests in test_database.py for comprehensive examples of how this all works.
 from collections import UserDict, defaultdict
 from dataclasses import dataclass
 
+from databuilder.itertools_utils import iter_flatten
 from databuilder.orm_utils import table_has_one_row_per_patient
-from tests.lib.util import iter_flatten
 
 
 class InMemoryDatabase:
