+++ tests/lib/databases.py
@@ -8,7 +8,7 @@ import sqlalchemy.exc
 from sqlalchemy.dialects import registry
 from sqlalchemy.orm import sessionmaker
 
-from .util import iter_flatten
+from databuilder.itertools_utils import iter_flatten
 
 MSSQL_SETUP_DIR = Path(__file__).parents[1].absolute() / "support/mssql"
