+++ tests/lib/util.py
@@ -8,15 +8,3 @@ class RecordingReporter:
 
 def null_reporter(msg):
     pass
-
-
-def iter_flatten(iterable, iter_classes=(list, tuple)):
-    """
-    Iterate over `iterable` recursively flattening any lists or tuples
-    encountered
-    """
-    for item in iterable:
-        if isinstance(item, iter_classes):
-            yield from iter_flatten(item, iter_classes)
-        else:
-            yield item
